+++ pylon_ros2_camera_component/include/internal/pylon_ros2_camera_impl.hpp
@@ -343,6 +343,8 @@ public:
 
     virtual std::string enablePTP(const bool& value) override;
 
+    virtual std::string getPTPStatus(int64_t& offset_from_master, std::string& status, std::string& servo_status);
+
     virtual std::string enableSyncFreeRunTimer(const bool& value) override;
 
     virtual std::string updateSyncFreeRunTimer() override;