+++ pylon_ros2_camera_component/include/pylon_ros2_camera.hpp
@@ -1091,6 +1091,12 @@ public:
     virtual std::string enablePTP(const bool& value) = 0;
 
     /**
+     * Sets the PTP priority - Applies to: ace 2 GigE.
+     * @return error message if an error occurred or done message otherwise.
+     */
+    virtual std::string getPTPStatus(int64_t& offset_from_master, std::string& status, std::string& servo_status) = 0;
+
+    /**
      * Enables the synchronous free run mode - Applies to: GigE and blaze.
      * @return error message if an error occurred or done message otherwise.
      */