+++ CHANGELOG.rst
@@ -2,9 +2,11 @@
 Changelog for package pylon_ros2_camera
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-2.2.1 (2024-10-25)
+2.3.0 (2024-10-25)
 -------------------
 * Taking into account auto exposure upper limit when user set is set to current setting.
+* Adding service to get ptp status.
+* Changing cmake minimum version.
 
 2.2.0 (2024-10-01)
 -------------------